v1.90:
- NRI: added C-compatible interface (macro magic based)
- NRI: dropped "static samplers" as a counter intuitive entity due to implementation in VK
- NRI: added "registerSpace" into "DescriptorSetDesc"
- NRI: added OMM support (WIP)
- NRI: fixed compilation issues with W4
- VK: expand VK wrapper with methods to receive proc addresses
- VK: fixed forgotten conservative raster enablement
- VK: removed unnecessary EXT postfix
- VK: added support for partially bound descriptors within descriptor set
- VK: fixed "free" after potential "return"
- VK: added support for sets starting not from "space = 0" in pipeline layout
- API: removed "noop" const
- API: renamed "baseSlot/setIndex" => "setIndexInPipelineLayout" to clarify what it is
- API: changed "CmdSetDescriptorSets" => "CmdSetDescriptorSet"
- API: fixed TextureTransition* helper funcs (WIP?)
- API: removed unused "DisplayDesc"
- API: removed NRIDescs.hpp file, functionality moved to NRIHelper.h (now C compatible)
- API: dropped bit fields (except "GeometryObjectInstance")
- API: added missing (or changed to minimal possible) "type" to "enum class"
- API: getting rid of Color<>
- README: added brief samples description
- simplified and improved scripts
- rewritten Cmake
- refactoring

+++ 4-Clean.bat
@@ -1,16 +1,9 @@
 @echo off
 
-set /P M=Do you want to delete PACKMAN repository? [y/n]
-if /I "%M%" neq "y" goto KEEP_PACKMAN
+if exist "build" rd /q /s "build"
 
-:DELETE_PACKMAN
-rd /q /s "%PM_PACKAGES_ROOT%"
-
-:KEEP_PACKMAN
-rd /q /s "_Build"
-rd /q /s "_Compiler"
-rd /q /s "_NRI_SDK"
-rd /q /s "External/AGS"
-rd /q /s "External/NVAPI"
-
-rd /q /s "build"
+if exist "_Build" rd /q /s "_Build"
+if exist "_Compiler" rd /q /s "_Compiler"
+if exist "_NRI_SDK" rd /q /s "_NRI_SDK"
+if exist "External/AGS" rd /q /s "External/AGS"
+if exist "External/NVAPI" rd /q /s "External/NVAPI"
